backend/requirements.txt
# Mini-C Compiler Backend Dependencies
# Compatible with Python 3.11+

# Web Framework
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0

# HTTP Client (required by openai)
httpx>=0.27.0,<1.0.0

# AI Integration
openai>=1.12.0,<2.0.0

# Data validation
pydantic>=2.5.0,<3.0.0

# Environment variables
python-dotenv>=1.0.0,<2.0.0